# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Feature Graph API — factory-managed feature databases and a knowledge graph over their entries
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface BuildInfo {
  version: string;
  environment: string;
  buildTime: string;
  /** @nullable */
  commit?: string | null;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
  guest: 'guest',
} as const;

export interface AuthUser {
  id: number;
  username: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthSession {
  user: AuthUser | null;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 40
     */
  username: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface FeatureTable {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  createdBy: string | null;
  createdAt: string;
}

export interface TableInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  category: string;
}

export type FieldDataType = typeof FieldDataType[keyof typeof FieldDataType];


export const FieldDataType = {
  string: 'string',
  number: 'number',
  boolean: 'boolean',
  date: 'date',
  json: 'json',
} as const;

export interface FeatureField {
  id: number;
  tableId: number;
  name: string;
  dataType: FieldDataType;
  /** @nullable */
  description?: string | null;
  required: boolean;
  createdAt: string;
}

export interface FieldInput {
  /** @minLength 1 */
  name: string;
  dataType: FieldDataType;
  description?: string;
  required?: boolean;
}

export type FeatureEntryData = { [key: string]: unknown };

export interface FeatureEntry {
  id: number;
  tableId: number;
  label: string;
  data: FeatureEntryData;
  createdAt: string;
  updatedAt: string;
}

export type EntryInputData = { [key: string]: unknown };

export interface EntryInput {
  /** @minLength 1 */
  label: string;
  data: EntryInputData;
}

export interface EntityRelation {
  id: number;
  fromEntryId: number;
  toEntryId: number;
  relationType: string;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  justification?: string | null;
  createdAt: string;
}

export interface EntryDetail {
  entry: FeatureEntry;
  table: FeatureTable;
  outgoing: EntityRelation[];
  incoming: EntityRelation[];
}

export interface RelationInput {
  fromEntryId: number;
  toEntryId: number;
  /** @minLength 1 */
  relationType: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  weight?: number;
  justification?: string;
}

export interface GraphNode {
  id: number;
  label: string;
  tableId: number;
  tableName: string;
  category: string;
}

export interface GraphEdge {
  id: number;
  source: number;
  target: number;
  relationType: string;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  justification?: string | null;
}

export interface GraphData {
  nodes: GraphNode[];
  edges: GraphEdge[];
}

export type GetGraphParams = {
search?: string;
tableId?: number;
limit?: number;
};