# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Feature Graph API — factory-managed feature databases and a knowledge graph over their entries
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthSession {
  user: AuthUser | null;
}